requirements.txt
# --- API & Server ---
fastapi
uvicorn
pydantic

# --- LLM & Agent Framework ---
langchain
langchain-core
langchain-community
langgraph
langchain-ollama

# --- Document Ingestion ---
docling

# --- Vector Database & Embeddings ---
qdrant-client
fastembed
sentence-transformers

# --- Evaluation & Observability ---
deepeval
opentelemetry-api
opentelemetry-sdk
openinference-instrumentation-langchain

# --- Session Persistence ---
redis
langgraph-checkpoint-redis

# front end
streamlit
requests